requirements.txt
fastapi
uvicorn
httpx
pillow
pyzbar
